requirements_free.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM Integration (FREE - Google Gemini, no card required!)
google-generativeai==0.3.2

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Frontend
streamlit==1.30.0

# Environment Management
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
